requirements.txt
pandas
numpy
matplotlib
scipy
scikit-learn
amazon-braket-sdk
